page 기능 추가 및 commentid int 변경

+++ src/main/java/com/aivle/bookapp/dto/CommentCreateRequest.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.Size;
 public class CommentCreateRequest {
 
     @NotNull(message = "책 ID는 필수입니다.")
-    private Long bookId;
+    private Integer bookId;
 
     @NotBlank(message = "댓글 내용은 필수입니다.")
     @Size(min = 1, max = 1000, message = "댓글은 1~1000자 이내로 입력해주세요.")
@@ -15,6 +15,6 @@ public class CommentCreateRequest {
 
     protected CommentCreateRequest() {}
 
-    public Long getBookId() { return bookId; }
+    public Integer getBookId() { return bookId; }
     public String getContent() { return content; }
 }
